V 0.6.3 Beta Build 45.8943 | Unstable
Added a highlight and Underline 2 with BRAND NEW SERIOUS BUGS. woohoo. they'll be in the issue tracker and should be fixed in the next build

Known issues:
The highlight and underline tools will select everything that has the same letters instead of just the thing you selected. this also includes images which change from an image to random letters